Folder of the game that I have so far
This may not be up to date. I used all the code from the latest commit.
I changed the name of RunnerGame4 to RunnerGame3 but it is the code for
RunnerGame4. I was too lazy to make a new class file so I copy pasted
and changed the class name. There is a txt file at the bottom that
details the changes I made.